Allow any strength cipher
I believe the cyrus default secprops come from a zeroed block of
memory, which importantly means that `max_ssf` = 0 so encryption is
actually disallowed by the client and server started by sasl_auth.

In the long term, it's better if the security parameters come from the
application but this patch sets up a reasonable and permissive default
which should be no worse than the previous situation of preventing
encryption altogether.

Note that encryption will also require wiring the sasl_encode and
sasl_decode interface functions.